REFACTOR: Relocate `mass_action_*` functions in catlog-wasm.
