+++ packages/catlog-wasm/src/theories.rs
@@ -15,76 +15,6 @@ use super::model_morphism::{MotifOccurrence, MotifsOptions, motifs};
 use super::result::JsResult;
 use super::{analyses::*, model::DblModel, theory::DblTheory};
 
-/// Simulates mass-action ODE on tabulated models.
-fn mass_action_tab(
-    model: &DblModel,
-    data: analyses::ode::MassActionProblemData,
-) -> Result<ODEResultWithEquations, String> {
-    let realised_model = model.discrete_tab()?;
-    let analysis = analyses::ode::StockFlowMassActionAnalysis::default();
-    let sys = analysis.build_system(realised_model, data.mass_conservation_type);
-    let sys_extended_scalars = analyses::ode::extend_mass_action_scalars(sys, &data);
-    let latex_equations = sys_extended_scalars
-        .map_variables(latex_ob_names_mass_action(model))
-        .to_latex_equations();
-    let analysis = analyses::ode::into_mass_action_analysis(sys_extended_scalars, data);
-    let solution = analysis.solve_with_defaults().map_err(|err| format!("{err:?}"));
-    Ok(ODEResultWithEquations {
-        solution: solution.into(),
-        latex_equations,
-    })
-}
-
-/// Simulates mass-action ODE on tabulated models.
-fn mass_action_modal(
-    model: &DblModel,
-    data: analyses::ode::MassActionProblemData,
-) -> Result<ODEResultWithEquations, String> {
-    let realised_model = model.modal()?;
-    let analysis = analyses::ode::PetriNetMassActionAnalysis::default();
-    let sys = analysis.build_system(realised_model, data.mass_conservation_type);
-    let sys_extended_scalars = analyses::ode::extend_mass_action_scalars(sys, &data);
-    let latex_equations = sys_extended_scalars
-        .map_variables(latex_ob_names_mass_action(model))
-        .to_latex_equations();
-    let analysis = analyses::ode::into_mass_action_analysis(sys_extended_scalars, data);
-    let solution = analysis.solve_with_defaults().map_err(|err| format!("{err:?}"));
-    Ok(ODEResultWithEquations {
-        solution: solution.into(),
-        latex_equations,
-    })
-}
-
-/// Generates mass-action equations for tabulated models.
-fn mass_action_equations_tab(
-    model: &DblModel,
-    mass_conservation_type: analyses::ode::MassConservationType,
-) -> Result<ODELatex, String> {
-    let realised_model = model.discrete_tab()?;
-    let analysis = analyses::ode::StockFlowMassActionAnalysis::default();
-    let sys = analysis.build_system(realised_model, mass_conservation_type);
-    let equations = sys
-        .map_variables(latex_ob_names_mass_action(model))
-        .extend_scalars(|param| param.map_variables(latex_mor_names_mass_action(model)))
-        .to_latex_equations();
-    Ok(ODELatex(equations))
-}
-
-/// Generates mass-action equations for tabulated models.
-fn mass_action_equations_modal(
-    model: &DblModel,
-    mass_conservation_type: analyses::ode::MassConservationType,
-) -> Result<ODELatex, String> {
-    let realised_model = model.discrete_tab()?;
-    let analysis = analyses::ode::StockFlowMassActionAnalysis::default();
-    let sys = analysis.build_system(realised_model, mass_conservation_type);
-    let equations = sys
-        .map_variables(latex_ob_names_mass_action(model))
-        .extend_scalars(|param| param.map_variables(latex_mor_names_mass_action(model)))
-        .to_latex_equations();
-    Ok(ODELatex(equations))
-}
-
 /// The empty or initial theory.
 #[wasm_bindgen]
 pub struct ThEmpty(Rc<theory::DiscreteDblTheory>);
